requirements.txt
openai
requests
PyMuPDF
python-dotenv
gdown
tiktoken
beautifulsoup4
feedparser
selenium
webdriver-manager
markdownify